+++ packages/popover/src/Popover.types.ts
@@ -142,6 +142,11 @@ export type PopoverProps = {
   className?: string;
 
   /**
+   * Class name applied to the popover content container
+   */
+  contentClassName?: string;
+
+  /**
    * Determines the alignment of the popover content relative to the trigger element
    *
    * default: `bottom`