+++ packages/testing-lib/src/index.ts
@@ -1,5 +1,10 @@
 import * as Context from './context';
 import * as jest from './jest';
 import * as JestDOM from './jest-dom';
+export { act, renderHook } from './RTLOverrides';
+export { waitForState } from './waitForState';
 
 export { Context, jest, JestDOM };
+
+export { eventContainingTargetValue } from './eventContainingTargetValue';
+export { tabNTimes } from './tabNTimes';